Move CustomVoxelShape to block package
