+++ src/main/java/com/mraof/minestuck/block/machine/AlchemiterBlock.java
@@ -1,9 +1,9 @@
 package com.mraof.minestuck.block.machine;
 
+import com.mraof.minestuck.block.CustomVoxelShape;
 import com.mraof.minestuck.block.EnumDowelType;
 import com.mraof.minestuck.block.MSProperties;
 import com.mraof.minestuck.blockentity.machine.AlchemiterBlockEntity;
-import com.mraof.minestuck.util.CustomVoxelShape;
 import com.mraof.minestuck.util.MSRotationUtil;
 import net.minecraft.MethodsReturnNonnullByDefault;
 import net.minecraft.core.BlockPos;
