+++ src/main/java/com/mraof/minestuck/block/machine/PunchDesignixBlock.java
@@ -1,8 +1,8 @@
 package com.mraof.minestuck.block.machine;
 
+import com.mraof.minestuck.block.CustomVoxelShape;
 import com.mraof.minestuck.block.MSProperties;
 import com.mraof.minestuck.blockentity.machine.PunchDesignixBlockEntity;
-import com.mraof.minestuck.util.CustomVoxelShape;
 import com.mraof.minestuck.util.MSRotationUtil;
 import net.minecraft.MethodsReturnNonnullByDefault;
 import net.minecraft.core.BlockPos;
